Rename values module to value_ops
